+++ examples/models/jax/gpt2/optimize.py
@@ -74,7 +74,7 @@ def main():
     dataloader = get_dataloader()
     verify_func = get_verify_function()
 
-    package = nav.jax.optimize(
+    package = nav.experimental.jax.optimize(
         model=model,
         model_params=params,
         dataloader=dataloader,