+++ model_navigator/experimental/jax/commands/export/exporters/__init__.py
@@ -1,5 +1,4 @@
-#!/usr/bin/env bash
-# Copyright (c) 2022, NVIDIA CORPORATION. All rights reserved.
+# Copyright (c) 2024, NVIDIA CORPORATION. All rights reserved.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,15 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from model_navigator.frameworks import is_jax_available, is_tf_available
 
-set -ex
-
-THIS_SCRIPT_PATH="$(realpath --relative-to="$(pwd)" "$0")"
-TEST_MODULE="$(dirname "${THIS_SCRIPT_PATH}"|sed 's/\//./g').test"
-
-pip install --upgrade "numpy<2" flax "jax[cuda12_pip]<0.4.16" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
-
-export XLA_PYTHON_CLIENT_PREALLOCATE=false
-python -m"${TEST_MODULE}" \
-    --status $(pwd)/status.yaml \
-    --verbose
+if is_tf_available() and is_jax_available():
+    from . import jax2savedmodel  # noqa: F401